+++ src/Koffee/MainModel.fs
@@ -268,12 +268,8 @@ with
         | Some host -> hist.WithNetHost host
         | None -> hist
 
-
-    static member private omitPathSortFromHistory sort =
-        sort = PathSort.Default
-
     member this.WithPathSort path sort =
-        if History.omitPathSortFromHistory sort then
+        if sort = PathSort.Default then
             { this with PathSort = this.PathSort.Remove path }
         else
             { this with PathSort = this.PathSort.Add(path, sort) }
@@ -430,6 +426,7 @@ type MainEvents =
     | Forward
     | ShowNavHistory
     | Refresh
+    | DeletePathSuggestion of Path
     | StartPrompt of PromptType
     | StartConfirm of ConfirmType
     | StartInput of InputType